Term UInt128

+++ Sources/IndrasNet/Instance/Term.swift
@@ -1,1 +1,1 @@
-typealias Term = Int
+typealias Term = UInt128
